Download CHELSA climate data

chelsa() downloads the CHELSA (Climatologies at high resolution for the earth’s land surface areas) V2.1 data of temperature and precipitation at a resolution of 30 arc sections.

chelsa(output_dir = NULL, var = "all", quiet = FALSE, ...)

Arguments

output_dir

Character (e.g., "../Desktop/chelsa"). Pathway to where the data will be stored.

var

Character. If supplied will download a subset of the climate data. Must be one of "all" (default), "prec", "tmax", "tmin" or "tave" to download the corresponding climate data.

quiet, ...

Arguments to control a download from the Internet download.file().

Value

Returns four subfolders named prec, tmax, tmin and tave. Each folder contains 12 GeoTiff (.tif) files, one for each month of the year for the time period 1981–2010. Each of the files are downloaded at a spatial resolution of 30 arc seconds (~1 km sq.). The precipitation folder contains average monthly precipitation (mm). The tmax folder contains maximum monthly temperature. The tmin folder contains minimum monthly temperature. The tave folder contains the average monthly temperature. The unit of measure for temperature is in °C.

When using the data, you agree to cite Karger et al. (2017, 2021).

References

Karger, D.N., Conrad, O., Bőhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W. et al (2017) Climatologies at high resolution for the earth’s land surface areas. Scientific Data, 4, 170122. doi:10.1038/sdata.2017.122

Karger, D.N., Conrad, O., Bőhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W. et al. (2021) Climatologies at high resolution for the earth’s land surface areas. EnviDat. doi:10.16904/envidat.228.v2.1

See also

Author

James L. Tsakalos and Martin R. Smith

Examples

# \donttest{

# Download time will depend on the size of the area you wish to access
# climate data for and your internet connection speed.

# Create temporary file
temp_path <- tempfile()

# Download the data
chelsa(
  output_dir = temp_path
)
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_01_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_01_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_02_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_02_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_03_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_03_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_04_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_04_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_05_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_05_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_06_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_06_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_07_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_07_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_08_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_08_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_09_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_09_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_10_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_10_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_11_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_11_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_12_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/pr/CHELSA_pr_12_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_01_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_01_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_02_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_02_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_03_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_03_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_04_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_04_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_05_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_05_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_06_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_06_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_07_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_07_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_08_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_08_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_09_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_09_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_10_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_10_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_11_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_11_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_12_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmin/CHELSA_tasmin_12_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_01_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_01_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_02_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_02_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_03_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_03_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_04_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_04_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_05_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_05_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_06_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_06_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_07_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_07_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_08_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_08_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_09_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_09_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_10_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_10_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_11_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_11_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_12_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tasmax/CHELSA_tasmax_12_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_01_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_01_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_02_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_02_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_03_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_03_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_04_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_04_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_05_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_05_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_06_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_06_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_07_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_07_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_08_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_08_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_09_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_09_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_10_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_10_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_11_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_11_1981-2010_V.2.1.tif'
#> Warning: downloaded length 0 != reported length 226
#> Warning: cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_12_1981-2010_V.2.1.tif': HTTP status was '404 Not Found'
#> Error in download.file(layer_url, file_path, mode = "wb", quiet = quiet,  : 
#>   cannot open URL 'https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/GLOBAL/climatologies/1981-2010/tas/CHELSA_tas_12_1981-2010_V.2.1.tif'

# Reset user options
unlink(file.path(temp_path))

# Note that unlike worldclim() we do not specify the location argument
# because it is not yet possible to extract smaller tile sections

# }