Download WorldClim climate data

worldclim() downloads the WorldClim V2.1 climate data for 1970-2000. This includes monthly climate data for minimum, mean, and maximum temperature and precipitation at a resolution of 0.5 minutes of a degree.

worldclim(output_dir, location, mode = "wb", var = "all", ...)

Arguments

output_dir

Character (e.g., "../Desktop/chelsa"). Pathway to where the data will be stored.

location

A "sp", "sf" polygon or point object.

var, mode, ...

Arguments to control a download from the Internet download.file().

Value

Creates four subfolders named prec, tmax, tmin and tmean. Each folder contains 12 GeoTiff (.tif) files, one for each month of the year for the time period 1970–2000. Each of the files are downloaded at a spatial resolution of 0.5 minutes of a degree. The precipitation folder contains average monthly precipitation (mm). The tmax folder contains maximum monthly temperature. The tmin folder contains minimum monthly temperature. The tmean folder contains the average monthly temperature. The unit of measure for temperature is in °C.

References

Fick, S.E. and R.J. Hijmans. (2017). WorldClim 2: new 1km spatial resolution climate surfaces for global land areas. International Journal of Climatology. 37, 4302–4315.

See also

Author

James L. Tsakalos

Examples

if (FALSE) {
# Download time will depend on the size of the area you wish to access
# climate data for and your internet connection speed.

# Import the Italian Biome data set
data("it_py", package = "climenv")

# Download the WorldClim data
worldclim(
  output_dir = "...Desktop/worldclim",
  location = "it_py"
)

}