Download elevation data

elev() downloads elevation data the Shuttle Radar Topography Mission (SRTM) , specifically the hole-filled CGIAR-SRTM (90 m resolution) for latitudes between -60 and 60 or Mapzen's synthesis digital elevation product.

elev(output_dir, loc, e_source = "mapzen")

Arguments

output_dir

Character (e.g., "../Desktop/chelsa"). Pathway to where the data will be stored.

loc

A "sp", "sf" polygon or point object.

e_source

Character. Elevation sourced from "mapzen" (default) or "geodata".

Value

Creates one subfolder named elev storing a raster (.tiff). If elevation is sourced from geodata the elevation is downloaded at a spatial resolution of 30 arc seconds (~1 km sq.). If elevation data is from mapzen then the product will be a mosaic. Specifically, Mapzen’s product is unique as it combines several sources of digital elevation models, including SRTM, the ArcticDEM (covering all areas north of 60°), EUDEM (digital elevation model over Europe; for review, see Mouratidis & Ampatzidis, 2019), and others into a single product. The resolution of this product was set to 7, corresponding to 611.5 m ground resolution at 60° latitude 864.8 m at 45° and 1223 m at 0°.

References

Hijmans, R.J., et al., (2023). geodata: Download Geographic Data. R package version 0.5-8. https://CRAN.R-project.org/package=geodataHollister, J. W., et al. (2021). elevatr: Access Elevation Data from Various APIs. R package version 0.4.2. https://CRAN.R-project.org/package=elevatrMouratidis, A., and Ampatzidis, D. (2019). European Digital Elevation Model Validation against Extensive Global Navigation Satellite Systems Data and Comparison with SRTM DEM and ASTER GDEM in Central Macedonia (Greece). ISPRS International Journal of Geo-Information 8, 108. DOI: 10.3390/ijgi8030108

See also

Author

James L. Tsakalos

Examples

if (FALSE) {
# Start by loading Italy's Mount Sibillini National Park boundary
data("Sibillini_py", package = "climenv")
# elevation will be saved in the output_dir (i.e. output directory)
   elev(output_dir = "...Desktop/elev", loc = Sibillini_py)
}